Fix #52. #53. Big change.
Read #52 for more details, but we're changing the API of each function
inside `src/` to receive `defaultFuncs` instead of `mergeWithDefaults`.
The former's an object that contains three functions:
- get
- post
- postFormData

They are equivalent to the ones in `utils` except they'll merge the
given form with extra attributes. Internally they call
`mergeWithDefaults`. We removed the ability to call the latter directly
because it's a statefull function, and can result in bugs later on (see
#52).