Cleaned up `getFriendsList`. Using `user_info_all` now.
We'd use to crawl someone's friend list slowly, because the UI would
only give us the friends in batches of 20 (and we needed to query them
sequentially). There's a new endpoint called `user_info_all` which
allows us to get all of the friends' info in one shot. It's way cleaner
and gives us more information (we were only returning friends IDs)
previously.